Refactor DefaultAnalyticsService for improved testability
- Moved processing steps to separate classes
- Introduced Analytics Processing Pipeline to process the various
Analytics steps in sequence
